# gh-118761: Speedup pathlib import by deferring shutil