--- a/Misc/NEWS.d/next/Library/2024-08-31-11-12-49.gh-issue-118761.Ai_Ma1.rst
+++ b/Misc/NEWS.d/next/Library/2024-08-31-11-12-49.gh-issue-118761.Ai_Ma1.rst
@@ -1,1 +1,0 @@
-Speed up pathlib import time by 15% by deferring shutil import.