~15% of pathlib import time (1.6ms on my machine) is spent on importing `shutil`. Since this module is only used in one place it makes sense to defer its import. Incidentally, this also makes the code more explicit imho, it took me a while to understand how the current code works.

![image](https://github.com/user-attachments/assets/e8c82258-45e7-4091-b554-4997574c19dc)


<!-- gh-issue-number: gh-118761 -->
* Issue: gh-118761
<!-- /gh-issue-number -->