@barneygale 2024-08-31
I think this probably shouldn't have a NEWS entry, because the `import shutil` was introduced only a couple of weeks ago, and there have been no releases from the `main` branch in that time. A user reading about a 15% performance improvement could be misled.