@hugovk 2024-08-31
> I think I used the PGO build, not sure what is the recommended build for benchmarking.

Ideally use [PGO+LTO](https://devguide.python.org/getting-started/setup-building/#optimization).

On macOS with PGO+LTO, running `./python.exe -X importtime -c "import pathlib"`, the pathlib import takes 7 ms with `main` and 6 ms with this PR.

You can see the big `shutil` block is missing underneath `pathlib._local` afterwards in the tuna importtime visualisations (I ran `./python.exe -c "import pathlib" && ./python.exe -X importtime -c "import pathlib" 2> import.log && tuna import.log`):

Before:

<img width="1582" alt="image" src="https://github.com/user-attachments/assets/0534139d-0cf2-45d0-9633-e198cbe28316">

After:

<img width="1582" alt="image" src="https://github.com/user-attachments/assets/5c3cd3e8-f450-4c51-ae99-c4765351feef">

</details>




And with hyperfine, this branch is about 1.5 ms faster than `main`:

```console
❯ hyperfine --warmup 32 \
--prepare "git checkout speedup-pathlib"  './python.exe -c "import pathlib"' \
--prepare "git checkout main"             './python.exe -c "import    pathlib"'
Benchmark 1: ./python.exe -c "import pathlib"
  Time (mean ± σ):      14.8 ms ±   0.6 ms    [User: 11.9 ms, System: 2.4 ms]
  Range (min … max):    14.1 ms …  18.2 ms    96 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs.

Benchmark 2: ./python.exe -c "import    pathlib"
  Time (mean ± σ):      16.4 ms ±   0.8 ms    [User: 13.1 ms, System: 2.7 ms]
  Range (min … max):    15.6 ms …  23.2 ms    91 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs.

Summary
  ./python.exe -c "import pathlib" ran
    1.11 ± 0.07 times faster than ./python.exe -c "import    pathlib"
```